==> paper_data_extraction/04_gene_extraction_experiment_b.ipynb <==
%load_ext autoreload
%autoreload 2

import json
import os
import pandas as pd
import papermill as pm
import scrapbook as sb

from functions import gpt

from tqdm.notebook import tqdm_notebook
tqdm_notebook().pandas()

pd.set_option('display.max_colwidth', None)

# project_folder
# sub_category
# work_bucket
# external_id
# base_name

input_file = f'truth_{external_id}.chunks.csv'
output_file = f'gpt_snps_{external_id}.csv'

data_path = os.getenv('DATA_PATH')
file_path = lambda *args: os.path.join(data_path, project_folder, sub_category, work_bucket,
                                       base_name, 'mathpix', *args)

model = 'gpt-4' # 'gpt-3.5-turbo' # 'gpt-4'
index_col = 'uuid'

system_prompt = """You are an academic researcher in information systems biology investigating possible links between DNA, 
RNA expression, and blood biomarkers for disease, especially aging and neurodegenerative pathologies 
(Alzheimer’s disease, Parkinson’s disease). 
- Please extract any SNP (Single Nucleotide Polymorphism) identifiers mentioned along with any relevant identifiers. 
- Please format your response in JSON. You only speak JSON. Do not write text that isn't JSON.
- If you do not know the answer return {}
"""

df = pd.read_csv(file_path(input_file), index_col=index_col)

json_errors = {}
general_errors = {}

def extract_genes(row):  
    response = None

    if len(row['text']) > 0:
        possible_response = gpt.chat_create(system_prompt, row['text'], model, output_json=True)
        if 'json_error' in possible_response:
            json_errors[row.name] = response
        elif 'general_error' in possible_response:
            general_errors[row.name] = response
        else:
            response = possible_response
        
    return response

df['snps'] = df.progress_apply(lambda row: extract_genes(row), axis=1)

with open(file_path('gpt_snps_json_errors.json'), 'w') as file:
    json.dump(json_errors, file)
    
with open(file_path('gpt_snps_general_errors.json'), 'w') as file:
    json.dump(general_errors, file)

df['snps'].to_csv(file_path(output_file))

sb.glue('status', 'completed')
